feat: add score api wrapper component
